+++ C++/Direct3D10/DeferredParticles/ParticleSystem.cpp
@@ -1,7 +1,8 @@
 //--------------------------------------------------------------------------------------
 // File: ParticleSystem.cpp
 //
-// Copyright (c) Microsoft Corporation. All rights reserved.
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License (MIT).
 //--------------------------------------------------------------------------------------
 #include "DXUT.h"
 #include "ParticleSystem.h"
@@ -330,7 +331,7 @@ void CParticleSystem::Init()
         m_pParticles[i].vPos += m_vCenter;
 
         m_pParticles[i].vDir.x = RPercent();
-        m_pParticles[i].vDir.y = fabs( RPercent() );
+        m_pParticles[i].vDir.y = fabsf( RPercent() );
         m_pParticles[i].vDir.z = RPercent();
         m_pParticles[i].vDir.x *= m_vDirMul.x;
         m_pParticles[i].vDir.y *= m_vDirMul.y;
@@ -365,9 +366,9 @@ void CParticleSystem::AdvanceSystem( float fTime, float fElapsedTime, D3DXVECTOR
         {
             float t = m_pParticles[i].fLife / m_fLifeSpan;
             float tm1 = t - 1.0f;
-            float fSizeLerp = 1.0f - pow( tm1, m_fSizeExponent );
-            float fSpeedLerp = 1.0f - pow( tm1, m_fSpeedExponent );
-            float fFadeLerp = 1.0f - pow( tm1, m_fFadeExponent );
+            float fSizeLerp = 1.0f - powf( tm1, m_fSizeExponent );
+            float fSpeedLerp = 1.0f - powf( tm1, m_fSpeedExponent );
+            float fFadeLerp = 1.0f - powf( tm1, m_fFadeExponent );
 
             float fSize = fSizeLerp * m_fEndSize + ( 1.0f - fSizeLerp ) * m_fStartSize;
             float fSpeed = fSpeedLerp * m_fEndSpeed + ( 1.0f - fSpeedLerp ) * m_fStartSpeed;
@@ -432,7 +433,7 @@ void CMushroomParticleSystem::Init()
         m_pParticles[i].vPos += m_vCenter;
 
         m_pParticles[i].vDir.x = RPercent();
-        m_pParticles[i].vDir.y = fabs( RPercent() );
+        m_pParticles[i].vDir.y = fabsf( RPercent() );
         m_pParticles[i].vDir.z = RPercent();
         m_pParticles[i].vDir.x *= m_vDirMul.x;
         m_pParticles[i].vDir.y *= m_vDirMul.y;
@@ -467,9 +468,9 @@ void CMushroomParticleSystem::AdvanceSystem( float fTime, float fElapsedTime, D3
         {
             float t = m_pParticles[i].fLife / m_fLifeSpan;
             float tm1 = t - 1.0f;
-            float fSizeLerp = 1.0f - pow( tm1, m_fSizeExponent );
-            float fSpeedLerp = 1.0f - pow( tm1, m_fSpeedExponent );
-            float fFadeLerp = 1.0f - pow( tm1, m_fFadeExponent );
+            float fSizeLerp = 1.0f - powf( tm1, m_fSizeExponent );
+            float fSpeedLerp = 1.0f - powf( tm1, m_fSpeedExponent );
+            float fFadeLerp = 1.0f - powf( tm1, m_fFadeExponent );
 
             float fSize = fSizeLerp * m_fEndSize + ( 1.0f - fSizeLerp ) * m_fStartSize;
             float fSpeed = fSpeedLerp * m_fEndSpeed + ( 1.0f - fSpeedLerp ) * m_fStartSpeed;
@@ -546,7 +547,7 @@ void CStalkParticleSystem::Init()
         m_pParticles[i].vPos += m_vCenter;
 
         m_pParticles[i].vDir.x = RPercent();
-        m_pParticles[i].vDir.y = fabs( RPercent() );
+        m_pParticles[i].vDir.y = fabsf( RPercent() );
         m_pParticles[i].vDir.z = RPercent();
         m_pParticles[i].vDir.x *= m_vDirMul.x;
         m_pParticles[i].vDir.y *= m_vDirMul.y;
@@ -581,9 +582,9 @@ void CStalkParticleSystem::AdvanceSystem( float fTime, float fElapsedTime, D3DXV
         {
             float t = m_pParticles[i].fLife / m_fLifeSpan;
             float tm1 = t - 1.0f;
-            float fSizeLerp = 1.0f - pow( tm1, m_fSizeExponent );
-            float fSpeedLerp = 1.0f - pow( tm1, m_fSpeedExponent );
-            float fFadeLerp = 1.0f - pow( tm1, m_fFadeExponent );
+            float fSizeLerp = 1.0f - powf( tm1, m_fSizeExponent );
+            float fSpeedLerp = 1.0f - powf( tm1, m_fSpeedExponent );
+            float fFadeLerp = 1.0f - powf( tm1, m_fFadeExponent );
 
             float fSize = fSizeLerp * m_fEndSize + ( 1.0f - fSizeLerp ) * m_fStartSize;
             float fSpeed = fSpeedLerp * m_fEndSpeed + ( 1.0f - fSpeedLerp ) * m_fStartSpeed;
@@ -724,9 +725,9 @@ void CGroundBurstParticleSystem::AdvanceSystem( float fTime, float fElapsedTime,
         {
             float t = m_pParticles[i].fLife / m_fLifeSpan;
             float tm1 = t - 1.0f;
-            float fSizeLerp = 1.0f - pow( tm1, m_fSizeExponent );
-            float fSpeedLerp = pow( tm1, m_fSpeedExponent );
-            float fFadeLerp = 1.0f - pow( tm1, m_fFadeExponent );
+            float fSizeLerp = 1.0f - powf( tm1, m_fSizeExponent );
+            float fSpeedLerp = powf( tm1, m_fSpeedExponent );
+            float fFadeLerp = 1.0f - powf( tm1, m_fFadeExponent );
             float fFade = fFadeLerp;
 
             D3DXVECTOR3 vDelta = m_pParticles[i].vPos - m_vCenter;
@@ -852,9 +853,9 @@ void CLandMineParticleSystem::AdvanceSystem( float fTime, float fElapsedTime, D3
         {
             float t = m_pParticles[i].fLife / m_fLifeSpan;
             float tm1 = t - 1.0f;
-            float fSizeLerp = 1.0f - pow( tm1, m_fSizeExponent );
-            float fSpeedLerp = pow( tm1, m_fSpeedExponent );
-            float fFadeLerp = 1.0f - pow( tm1, m_fFadeExponent );
+            float fSizeLerp = 1.0f - powf( tm1, m_fSizeExponent );
+            float fSpeedLerp = powf( tm1, m_fSpeedExponent );
+            float fFadeLerp = 1.0f - powf( tm1, m_fFadeExponent );
 
             float fFade = fFadeLerp;
