+++ C++/Direct3D10/GPUSpectrogram/AudioData.cpp
@@ -1,7 +1,8 @@
 //--------------------------------------------------------------------------------------
 // File: AudioData.cpp
 //
-// Copyright (c) Microsoft Corporation. All rights reserved.
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License (MIT).
 //--------------------------------------------------------------------------------------
 #include "AudioData.h"
 #include <mmsystem.h>
@@ -300,7 +301,7 @@ bool CAudioData::NormalizeData()
         for( unsigned long i = 0; i < m_ulNumSamples; i++ )
         {
             if( fabs( m_ppChannel[ ulChannel ][i] ) > fSampleMax )
-                fSampleMax = fabs( m_ppChannel[ ulChannel ][i] );
+                fSampleMax = fabsf( m_ppChannel[ ulChannel ][i] );
         }
 
         // Normalize
